start guessing wallabag version
On testing the connection, differentiate between v1 and v2. Testing
connection with v2 works, but v1 is not yet reimplemented.
