+++ app/src/main/java/fr/gaulupeau/apps/Poche/data/FeedsCredentials.java
@@ -1,0 +1,9 @@
+package fr.gaulupeau.apps.Poche.data;
+
+/**
+ * Created by strubbl on 11.04.16.
+ */
+public class FeedsCredentials {
+    public String userID;
+    public String token;
+}
