+++ app/src/main/java/fr/gaulupeau/apps/Poche/data/Settings.java
@@ -24,6 +24,8 @@ public class Settings {
     public static final String HTTP_AUTH_PASSWORD = "http_auth_password";
     public static final String THEME = "theme";
     public static final String CONFIGURE_OPTIONAL_DIALOG_SHOWN = "configure_optional_dialog_shown";
+    public static final String WALLABAG_LOGIN_FORM_V1 = "<form method=\"post\" action=\"?login\" name=\"loginform\">";
+    public static final String WALLABAG_LOGIN_FORM_V2 = "<form action=\"/login_check\" method=\"post\" name=\"loginform\">";
 
     private SharedPreferences pref;
